sql_agent/requirements.txt
openai
psycopg2-binary
python-dotenv
fastapi
uvicorn
pydantic
